create script to sync time with nic time servers
refactor code
